Move abode base entities to separate module
